requirements.txt
matplotlib==3.6.3
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
plotly==5.9.0
scikit_learn==1.2.1
seaborn==0.11.2
streamlit==1.17.0
xgboost==1.7.3
black==22.6.0
isort==5.9.3
flake8==4.0.1
statsmodels==0.13.2